Added fallback to English in Journal
Previously it would just not change the image if Journal couldn't find
an image file for some language. This fallback does exist on Windows
version, but for some reason wasn't implemented for Unix one.
I also added some messages to clearly tell what's happening and why it
didn't use the locale that was specified in the message.